refactor(browsing): centralize context routing contracts
